Reorganize the code a bit, and fix bugs
