+++ sys/net/gnrc/transport_layer/tcp_freebsd/Makefile
@@ -2,5 +2,6 @@ MODULE = gnrc_tcp_freebsd
 
 DIRS += blip
 DIRS += bsdtcp
+DIRS += lib
 
 include $(RIOTBASE)/Makefile.base
