+++ sys/net/gnrc/transport_layer/tcp_freebsd/bsdtcp/tcp_input.c
@@ -79,8 +79,8 @@
 #include "tcp_seq.h"
 #include "tcp_timer.h"
 #include "tcp_var.h"
-#include "bitmap.h"
-#include "cbuf.h"
+#include "../lib/bitmap.h"
+#include "../lib/cbuf.h"
 #include "icmp_var.h"
 #include "ip.h"
 #include "ip6.h"
