+++ sys/net/gnrc/transport_layer/tcp_freebsd/bsdtcp/tcp_timer.c
@@ -34,7 +34,7 @@
 #include <stdio.h>
 
 #include "../gnrc_tcp_freebsd_internal.h"
-#include "lbuf.h"
+#include "../lib/lbuf.h"
 #include "tcp_fsm.h"
 #include "tcp_timer.h"
 #include "tcp_var.h"
