+++ sys/net/gnrc/transport_layer/tcp_freebsd/lib/Makefile
@@ -1,0 +1,3 @@
+MODULE = gnrc_tcp_freebsd_internal
+
+include $(RIOTBASE)/Makefile.base
