+++ sys/net/gnrc/transport_layer/tcp_freebsd/lib/bitmap.h
